Read data files; change to CartoCrow
This (re-)adds the functionality for the user to upload a data file,
instead of having to manually fill in the data fields. As opposed to the
earlier implementation, the user now has to click the Import button to
upload the actual data file, and then the data values from it are
imported into the data fields. This way, the user can still make
adjustments if desired.

This moves the map files to wwwroot/data so that they are accessible and
can be requested by the JavaScript code.

This commit also changes the name GeoViz to CartoCrow, and introduces
some CartoCrow branding throughout the interface.

Finally, this commit adds comments to the JavaScript methods.