Fix undefined behavior by iterator invalidation
Modifying an std::vector invalidates its iterators, hence causing
undefined behavior in this for-loop. This is undefined behavior even
if the vector is preallocated by calling reserve() (as was done here).

This results in an error when building in debug mode with MSVC. Fixed
by manually iterating over the index.

[1] https://stackoverflow.com/questions/35468207