Fix out-of-bounds strcmp
Obvious bug because of a missing \0 at the end of the string. Apparently
this works fine in practice under g++/clang, but it crashes under MSVC.

On the long term, we shouldn't be using strcmp at all, but let's keep
that for later. I performed the minimal fix for now.

Also added some useful debug output.