Render necklace maps with the GeometryRenderer
The aim is to eventually get rid of the 1000+ lines of code generating
SVG code, and instead do this generically using the GeometryRenderer.
This will enable us to output the necklace map to Ipe rather than SVG,
and to show it interactively using the GeometryWidget.

This is still very incomplete, because on one hand the GeometryRenderer
doesn't support many features we need yet, and on the other hand, we
would need to parse SVG styles etc. from the input SVG to be able to use
the same style with the GeometryRenderer. (The original necklace map
code just copies the input SVG style string to the output SVG shapes,
which works, but doesn't allow us to find the color of a region easily,
for example.) It seems silly to parse SVG style strings, so instead I'd
like to just change the input to an Ipe file (which is very easy to
parse).