Initial version of using Ipelib to read maps
We can now read an Ipe file and see which shapes are in there. We cannot
yet read the actual shapes, so instead we generate some dummy shapes for
now.

I have no idea if this compiles under Windows, but probably Ipelib
(which is not really made for being compiled under MSVC) is going to be
slightly problematic to get to work.