Start big refactor of the CartoCrow core
The intention is to make it possible to CGAL's exact computation mode
(currently not supported by CartoCrow because we use floating-point
number operations all over the place). Exact computation will not be
possible for necklace maps and flow maps (and also is not really
necessary there) but we do want it for other algorithms to be
implemented later (e.g., simplification).

This implies that all kinds of functionality that relies on
floating-point computation has to be removed from the core and instead
moved to necklace / flow maps, where it belongs. The core will then
contain:

* aliases for basic CGAL types;
* functions for Ipe import / export;
* maybe functions for coordinate transforms, map projections, etc.?
* other utilities such as a timer.

This commit makes a start with moving things around, and contains a
refactored Timer class to make it more usable. Currently this does not
even compile, let alone work.