Change the sweep circle to be actually circular
Intervals are now not split anymore at the φ = π branch cut. Instead
intervals can simply span over the branch cut.

Of course, there still is a branch cut that needs to be handled, if the
circle grows over an intersection of some SweepEdge with the branch cut.
This is now also implemented.

Also some other bugs have been fixed. JoinEvents now store weak_ptrs to
their edges, and if the weak_ptrs expire, the event is considered
invalid and not handled.

This means that in theory the entire algorithm should work now even for
edges spanning the branch cut. Unfortunately there seems to be at least
one other bug that causes the sweep circle to become invalid, causing
crashes.