Fix floating-point inaccuracies in phiForR()
The computation for segments would ask PolarSegment for intersections
with a circle of radius r, but due to floating-point inaccuracies, if r
is very close to the near or far endpoint of the segment, sometimes no
intersection is found. This commit just takes the φ of the near or far
endpoint in this case instead of... crashing.