Fix handling of edges crossing over φ = π
Edges should never cross, so the code handling φ = π didn't handle this
situation. However, if a join event happens exactly on φ = π, then due
to floating-point issues (that is, the binary search finding the join
event not being 100% accurate) it can happen that they intersect just
before the join event, which then breaks the φ = π handling, making the
sweep circle invalid.

This commit fixes this issue by first figuring out which edges need to
be reinserted both for counter-clockwise and clockwise wraparounds, and
only then reinserting all of them.