Fix missing join events after join events
If we handle a join event, we have to insert new join events as well.
This commit moves handleJoinEvents() from VertexEvent to Event to
facilitate this, and then calls it from JoinEvent::handle().

One detail is that handleJoinEvents() works by taking the φ of the
event, and looking for edges there. For vertex events this always works
as the edges and spirals inserted for them have nearR == currentR. So
then we can always exactly find these at the exact φ. For join events
this is not necessarily the case due to floating-point inaccuracies: the
φ of both joined edges can differ slightly.

To solve this, we now update the join event's φ value to the exact φ of
the remaining edge (if any). This way, insertJoinEvents() can always
find the remaining edge correctly.