Solve FP robustness issues due to pruning
Pruning a sweep edge can just slightly change its φ value for a given r,
due to rounding errors. This presents various problems.

Firstly, if I prune an edge, I may not be able to find that edge anymore
in the sweep circle. Namely, I changed the edge's shape, but the edge is
still indexed with the old shape as key in the std::multimap that stores
the sweep circle edges. If these yield different φ values, I've lost the
edge. This commit solves this problem by using an std::multiset to store
the sweep circle instead. This way, the φ comparisons are done using the
shapes stored in the edges themselves, rather than the shape stored as a
key separately.

Secondly, pruning edges before their join event is nice because then for
sure the endpoints of the edges have the same φ, hence erasing them from
the sweep circle always works without problems. On the other hand, doing
it this way also presents a new problem, namely that if we prune, we may
move the endpoint over φ = π, and now the sweep circle's order is messed
up. So we don't actually want to do it this way: we now first remove the
edges from the sweep circle, and then we prune. This seems to just work,
at least under g++/clang, but on MSVC it seems that the removal may miss
the edges. This commit solves this problem by adding an additional check
to the edge comparator used by the sweep circle: if the edge that we are
checking actually has a lower φ than its previous neighbor edge (that we
luckily already store as a pointer), then we sneakily compare not to the
edge's own φ but the previous edge's φ. To make this work, some annoying
special cases have to be handled (e.g., if the edge hasn't been inserted
yet it doesn't yet have a previous edge, and if it has been inserted but
it is the first edge on the sweep circle, we shouldn't change its φ, and
so on).