Make join events invalid if their edges are gone
This already used to work because the std::shared_ptr<SweepEdge>
carrying the join event's sweep edges would become stale. However, now
we keep references to these shared_ptrs because we need them for the
second sweep. Hence, we need to explicitly maintain if a SweepEdge is
still on the circle. This commit implements just that.