Fix left vertex events to split reachable regions
A left or right vertex event produces a reachable region, but not the
entire region is reachable from the same node: part is reachable from
the newly inserted node (for the vertex), while part inherits its
reachable node from the existing reachable region. This commit
implements this.

To do: implement the same for right vertex events, and properly remove
all reachable regions belonging to a leaf node -- not just the
incident ones -- in join events.